net/mlx5: Keep buffer in FW ownership for manual cable length input
Update the behavior of buffer ownership when the cable length is
manually set by the user. Previously, the driver stored the cable
length as a variable and transitioned the buffer to SW ownership when
the user provided cable length, becoming responsible for calculating the
xoff and xon thresholds.

With this change, the cable length is now stored directly in the device.
When the user provides a cable length, the buffer remains in FW
ownership, and the FW is responsible for calculating the xoff and xon
thresholds.

Issue: 4163956
Change-Id: Ieadb8ec352f9bde56dbdbbcce96f0bb5cce080f7
Signed-off-by: Oren Sidi <osidi@nvidia.com>
Reviewed-by: Tariq Toukan <tariqt@nvidia.com>
Signed-off-by: Alexei Lazar <alazar@nvidia.com>