net/mlx5: Store QoS sched nodes in the sh_devlink
In order to support cross-esw scheduling, the nodes for all devices need
to be accessible from the same place so that normalization can work.
Store them in the shared devlink instance and provide an accessor for
them.

Protecting against concurrent QoS modifications is now done with
the shd lock. All calls originating from devlink rate already have the
shd lock held so only the additional entry points into QoS need to
acquire it.

As part of this change, the E-Switch qos domain concept was removed.
E-Switch QoS domains were added with the intention of eventually
implementing shared qos domains to support cross-esw scheduling in the
previous approach ([1]), but they are no longer necessary in the new
approach.

[1] https://lore.kernel.org/netdev/20250213180134.323929-1-tariqt@nvidia.com/

Issue: 3645895
Signed-off-by: Cosmin Ratiu <cratiu@nvidia.com>
Change-Id: Ic45a16b87743ef42b8617275dc1fa8f294339eaa
Reviewed-by: Carolina Jubran <cjubran@nvidia.com>
Signed-off-by: Alexei Lazar <alazar@nvidia.com>