net/mlx5e: Drop unused channel parameters
The channel parameters from struct mlx5_qmgmt_data are
built in mlx5e_queue_mem_alloc() but are not used.

mlx5e_open_channel() builds the channel parameters internally and those
parameters will be the ones that are used when opening the queue.

This patch drops the unused parameters.

Issue: 4466747
Signed-off-by: Dragos Tatulea <dtatulea@nvidia.com>
Change-Id: Id2f91f3d77ed510b05bff72a27ffcff41ba6c3c5
Reviewed-by: Tariq Toukan <tariqt@nvidia.com>
Signed-off-by: Alexei Lazar <alazar@nvidia.com>